third_party/blink/web_tests/fast/forms/select/listbox-scroll-to-selection-expected.txt

List boxes should be scrolled to show the selection when they are added but shouldn't scroll parent elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.scrollingElement.scrollTop is 0
PASS document.getElementById("test1").scrollTop is > 0
PASS successfullyParsed is true

TEST COMPLETE

foo
foo
foo
foo
foo
bar
